requirements.txt
streamlit==1.45.0
openai==1.77.0
langchain==0.3.25
pillow==10.1.0
pandas==2.2.3
requests==2.31.0
inquirer==3.1.3
python-dotenv==1.0.0